pyproject.toml
[tool.poetry]
name = "eob-reporting"
version = "0.1.0"
description = ""
authors = ["DimaSimakov <dimarodeo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
certifi = "^2024.2.2"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
openpyxl = "^3.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
